Improve #split_message behavior (#662)
Previously, it would just split at 2k characters, without taking into account if that would break a word in half, or possible term a valid url/resolvable link, into a broken one. Now it splits the message on the last space character or keeps it the same if there are just no space characters at all.